+++ sumo/cli/dosplot.py
@@ -43,6 +43,7 @@ def dosplot(filename=None, prefix=None, directory=None, elements=None,
             legend_frame_on=False, legend_cutoff=3., gaussian=None, height=6.,
             width=8., xmin=-6., xmax=6., num_columns=2, colours=None, yscale=1,
             xlabel='Energy (eV)', ylabel='Arb. units',
+            style=None, no_base_style=False,
             image_format='pdf', dpi=400, plt=None, fonts=None):
     """A script to plot the density of states from a vasprun.xml file.
 
@@ -128,6 +129,10 @@ def dosplot(filename=None, prefix=None, directory=None, elements=None,
         xlabel (:obj:`str`, optional): Label/units for x-axis (i.e. energy)
         ylabel (:obj:`str`, optional): Label/units for y-axis (i.e. DOS)
         yscale (:obj:`float`, optional): Scaling factor for the y-axis.
+        style (:obj:`list` or :obj:`str`, optional): (List of) matplotlib style
+            specifications, to be composed on top of Sumo base style.
+        no_base_style (:obj:`bool`, optional): Prevent use of sumo base style.
+            This can make alternative styles behave more predictably.
         image_format (:obj:`str`, optional): The image file format. Can be any
             format supported by matplotlib, including: png, jpg, pdf, and svg.
             Defaults to pdf.
@@ -164,7 +169,8 @@ def dosplot(filename=None, prefix=None, directory=None, elements=None,
                            legend_frame_on=legend_frame_on,
                            xlabel=xlabel, ylabel=ylabel,
                            legend_cutoff=legend_cutoff, dpi=dpi, plt=plt,
-                           fonts=fonts)
+                           fonts=fonts, style=style,
+                           no_base_style=no_base_style)
 
     if save_files:
         basename = 'dos.{}'.format(image_format)
@@ -271,9 +277,9 @@ def _get_parser():
                         help='only plot the total density of states')
     parser.add_argument('--no-total', action='store_false', dest='total',
                         help='don\'t plot the total density of states')
-    parser.add_argument('--height', type=float, default=6.,
+    parser.add_argument('--height', type=float, default=None,
                         help='height of the graph')
-    parser.add_argument('--width', type=float, default=8.,
+    parser.add_argument('--width', type=float, default=None,
                         help='width of the graph')
     parser.add_argument('--xmin', type=float, default=-6.,
                         help='minimum energy on the x-axis')
@@ -281,10 +287,15 @@ def _get_parser():
                         help='maximum energy on the x-axis')
     parser.add_argument('--config', type=str, default=None,
                         help='colour configuration file')
+    parser.add_argument('--style', type=str, nargs='+', default=None,
+                        help='matplotlib style specifications')
+    parser.add_argument('--no-base-style', action='store_true',
+                        dest='no_base_style',
+                        help='prevent use of sumo base style')
     parser.add_argument('--xlabel', type=str, default='Energy (eV)',
-                        help='x-axis (i.e. energy) label/units'),
+                        help='x-axis (i.e. energy) label/units')
     parser.add_argument('--ylabel', type=str, default='Arb. units',
-                        help='y-axis (i.e. DOS) label/units'),
+                        help='y-axis (i.e. DOS) label/units')
     parser.add_argument('--yscale', type=float, default=1,
                         help='scaling factor for the y-axis')
     parser.add_argument('--format', type=str, default='pdf',
@@ -328,9 +339,10 @@ def main():
             legend_cutoff=args.legend_cutoff, gaussian=args.gaussian,
             height=args.height, width=args.width, xmin=args.xmin,
             xmax=args.xmax, num_columns=args.columns, colours=colours,
+            style=args.style, no_base_style=args.no_base_style,
             xlabel=args.xlabel, ylabel=args.ylabel,
             yscale=args.yscale, image_format=args.image_format, dpi=args.dpi,
-            fonts=[args.font])
+            fonts=args.font)
 
 
 if __name__ == "__main__":